Add configurable startpage
When the user access the uri / we want to display the content of page "home".
This can be configured with the config option `startpage`. This is completely
handled by the page repository. If we request the page for `uri` "/" a page with
this $uri and the content of page "home" is returned.